Support constructor Haddocks in more places
Summary:
This adds support for adding Haddocks on individual non-record fields
of regular (and GADT) constructors. The following now parses just fine
with `-haddock` enabled:

    data Foo
      = Baz             -- ^ doc on the `Baz` constructor
          Int           -- ^ doc on the `Int` field of `Baz`
          String        -- ^ doc on the `String` field of `Baz`

      | Int             -- ^ doc on the `Int` field of the `:*` constructor
          :*            -- ^ doc on the `:*` constructor
        String          -- ^ doc on the `String` field of the `:*` constructor

      | Boa             -- ^ doc on the `Boa` record constructor
          { y :: () }

The change is backwards compatible: if there is only one doc and it occurs
on the last field, it is lifted to apply to the whole constructor (as
before).

Test Plan: None yet

Reviewers: austin, bgamari, alanz

Subscribers: rwbarton, thomie, mpickering

GHC Trac Issues: #8822

Differential Revision: https://phabricator.haskell.org/D4094